move game logic files into their own module
