+++ src/game/mod.rs
@@ -1,0 +1,13 @@
+mod actions;
+mod audio;
+mod finish_areas;
+mod game_object;
+mod player;
+mod collision;
+
+pub use actions::ActionsPlugin;
+pub use audio::InternalAudioPlugin;
+pub use finish_areas::FinishAreaPlugin;
+pub use game_object::GameObjectPlugin;
+pub use player::PlayerPlugin;
+
